\name{Q}
\alias{Q}
\title{
Quantile
}
\description{
Calculates the quantile
}
\usage{
Q(par, j, n)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{par}{Matrix of parameters (nmeric 4xK - matrix); may also be delivered separately.}
  \item{j}{Possible values (numeric vector of length J > 0)}
  \item{n}{Frequencies (integer vector of length J; non-negative entries)}
}

\author{
Lisa Allmesberger
Fabian Bergs
Stefan Immler
Michael Kässmann
}





