\name{Q}
\alias{Q}
\title{
Quantile
}
\description{
Calculates the quantile
}
\usage{
Q(par, j, n)
}

\arguments{
  \item{par}{Matrix of parameters (nmeric Kx4 - matrix); may also be delivered separately.}
  \item{j}{Unique non-resistant observations (numeric vector of length J > 0)}
  \item{n}{Frequencies (integer vector of length J; non-negative entries)}
}

\author{
Lisa Allmesberger
Fabian Bergs
Stefan Immler
Michael Kässmann
}






