\name{Q3}
\alias{Q3}
\title{
Quantile3
}
\description{
Quantile 3
}
\usage{
Q3(s, a, b)
}

\arguments{
  \item{s}{Logarithm of standard deviation}
  \item{a}{Parameters of invers Gamma distribution}
  \item{b}{Parameters of invers Gamma distribution}
}

\author{
Lisa Allmesberger
Fabian Bergs
Stefan Immler
Michael Kässmann
}





