\name{Qmax}
\alias{Qmax}
\title{
Quantile Max
}
\description{
Quantile Max
}
\usage{
Qmax(par, tau, j, n, a, b)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{par}{Matrix of parameters (nmeric 4xK - matrix); may also be delivered separately.}
  \item{tau}{A value.}
  \item{j}{Possible values (numeric vector of length J > 0)}
  \item{n}{Frequencies (integer vector of length J; non-negative entries)}
  \item{a}{a}
  \item{b}{a}
}

\author{
Lisa Allmesberger
Fabian Bergs
Stefan Immler
Michael Kässmann
}

